Height queries in compute shader

+++ crest/Assets/Crest/Crest/Shaders/OceanLODData.hlsl
@@ -53,7 +53,7 @@ float2 LD_WorldToUV(in float2 i_samplePos, in float2 i_centerPos, in float i_res
 	return (i_samplePos - i_centerPos) / (i_texelSize * i_res) + 0.5;
 }
 
-float3 WorldToUV(in float2 i_samplePos, in float i_sliceIndex) {
+float3 WorldToUV(in float2 i_samplePos, in uint i_sliceIndex) {
 	const float2 result = LD_WorldToUV(
 		i_samplePos,
 		_LD_Pos_Scale[i_sliceIndex].xy,
@@ -63,7 +63,7 @@ float3 WorldToUV(in float2 i_samplePos, in float i_sliceIndex) {
 	return float3(result, i_sliceIndex);
 }
 
-float3 WorldToUV_BiggerLod(in float2 i_samplePos, in float i_sliceIndex_BiggerLod) {
+float3 WorldToUV_BiggerLod(in float2 i_samplePos, in uint i_sliceIndex_BiggerLod) {
 	const float2 result = LD_WorldToUV(
 		i_samplePos, _LD_Pos_Scale[i_sliceIndex_BiggerLod].xy,
 		_LD_Params[i_sliceIndex_BiggerLod].y,
@@ -72,7 +72,7 @@ float3 WorldToUV_BiggerLod(in float2 i_samplePos, in float i_sliceIndex_BiggerLo
 	return float3(result, i_sliceIndex_BiggerLod);
 }
 
-float3 WorldToUV_Source(in float2 i_samplePos, in float i_sliceIndex_Source) {
+float3 WorldToUV_Source(in float2 i_samplePos, in uint i_sliceIndex_Source) {
 	const float2 result = LD_WorldToUV(
 		i_samplePos,
 		_LD_Pos_Scale_Source[i_sliceIndex_Source].xy,
@@ -100,6 +100,7 @@ float2 IDtoUV(in float2 i_id, in float i_width, in float i_height)
 {
 	return (i_id + 0.5) / float2(i_width, i_height);
 }
+
 
 // Sampling functions
 void SampleDisplacements(in Texture2DArray i_dispSampler, in float3 i_uv_slice, in float i_wt, inout float3 io_worldPos, inout float io_sss)
